% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseMagnetic.R
\name{analyseMagnetic}
\alias{analyseMagnetic}
\title{Analyse geomagnetic data from archaeological context}
\usage{
analyseMagnetic(anomalies_sdf, magnetic_raster, searchradius = 2,
  dipolfactor = 2, angle_steps = 10, get_profile_values = FALSE,
  get_dipol = FALSE, method = "avg", cut_value = 1.5, dipol_minima = -4)
}
\arguments{
\item{anomalies_sdf}{Spatialdataframe: Point data containing the Anomalies}

\item{magnetic_raster}{RasterLayer: Geomagnetic data (nT)}

\item{searchradius}{The Radius around every anomaly that is under observation}

\item{dipolfactor}{1. version of defining a dipole: The factor used to descripe a dipole; A dipol is defined by |minimal value * dipolfactor| > maximal value}

\item{angle_steps}{Profiles are generated around the anomalie points in steps of angle_steps degrees}

\item{get_profile_values}{Calculating the values of all profiles (TRUE)}

\item{get_dipol}{Identifing the dipoles (TRUE)}

\item{method}{method for calculating the profile values  avg = Average width of all profiles; median = Median width of all profiles}

\item{cut_value}{nT-value at which the width of the amplitude will be determined}

\item{dipol_minima}{Additional defining of a dipole: All anomalies with min nT < dipol_minima in the profile are defined as dipoles}
}
\value{
A spatial data frame with the results of the dipole identifing (0/1) or/and the width and heigth of the amplitudes
}
\description{
Analyse geomagnetic data from archaeological context
}
\examples{
sum(1:10)
}
